requirements.txt
streamlit
seleniumbase
playwright
httpx
sqlalchemy
apscheduler
google-generativeai
bcrypt
pybase64
python-dotenv
structlog
pandas
openpyxl
pillow
tenacity
aiofiles
plotly
requests
beautifulsoup4
fake-useragent
cryptography
